+++ src/com/maddyhome/idea/vim/group/WindowGroup.java
@@ -22,12 +22,17 @@ import com.intellij.openapi.actionSystem.DataContext;
 import com.intellij.openapi.actionSystem.PlatformDataKeys;
 import com.intellij.openapi.fileEditor.ex.FileEditorManagerEx;
 import com.intellij.openapi.fileEditor.impl.EditorWindow;
+import com.intellij.openapi.fileEditor.impl.EditorWithProviderComposite;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.maddyhome.idea.vim.VimPlugin;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
+import java.awt.*;
+import java.util.*;
+import java.util.List;
 
 public class WindowGroup {
   public WindowGroup() {
@@ -83,7 +88,60 @@ public class WindowGroup {
     splitWindow(SwingConstants.VERTICAL, context, filename);
   }
 
-  private FileEditorManagerEx getFileEditorManager(@NotNull DataContext context) {
+  public void selectWindowInRow(@NotNull DataContext context, int relativePosition, boolean vertical) {
+    final FileEditorManagerEx fileEditorManager = getFileEditorManager(context);
+    final EditorWindow currentWindow = fileEditorManager.getCurrentWindow();
+    if (currentWindow != null) {
+      final EditorWindow[] windows = fileEditorManager.getWindows();
+      final List<EditorWindow> row = findWindowsInRow(currentWindow, Arrays.asList(windows), vertical);
+      selectWindow(currentWindow, row, relativePosition);
+    }
+  }
+
+  private void selectWindow(@NotNull EditorWindow currentWindow, @NotNull List<EditorWindow> windows,
+                            int relativePosition) {
+    final int pos = windows.indexOf(currentWindow);
+    final int selected = pos + relativePosition;
+    final int normalized = Math.max(0, Math.min(selected, windows.size() - 1));
+    windows.get(normalized).setAsCurrentWindow(true);
+  }
+
+  @NotNull
+  private static List<EditorWindow> findWindowsInRow(@NotNull EditorWindow anchor,
+                                                     @NotNull List<EditorWindow> windows, final boolean vertical) {
+    final Rectangle anchorRect = getEditorWindowRectangle(anchor);
+    if (anchorRect != null) {
+      final List<EditorWindow> result = new ArrayList<EditorWindow>();
+      final double coord = vertical ? anchorRect.getX() : anchorRect.getY();
+      for (EditorWindow window : windows) {
+        final Rectangle rect = getEditorWindowRectangle(window);
+        if (rect != null) {
+          final double min = vertical ? rect.getX() : rect.getY();
+          final double max = min + (vertical ? rect.getWidth() : rect.getHeight());
+          if (coord >= min && coord <= max) {
+            result.add(window);
+          }
+        }
+      }
+      Collections.sort(result, new Comparator<EditorWindow>() {
+        @Override
+        public int compare(EditorWindow window1, EditorWindow window2) {
+          final Rectangle rect1 = getEditorWindowRectangle(window1);
+          final Rectangle rect2 = getEditorWindowRectangle(window2);
+          if (rect1 != null && rect2 != null) {
+            final double diff = vertical ? (rect1.getY() - rect2.getY()) : (rect1.getX() - rect2.getX());
+            return diff < 0 ? -1 : diff > 0 ? 1 : 0;
+          }
+          return 0;
+        }
+      });
+      return result;
+    }
+    return Collections.singletonList(anchor);
+  }
+
+  @NotNull
+  private static FileEditorManagerEx getFileEditorManager(@NotNull DataContext context) {
     final Project project = PlatformDataKeys.PROJECT.getData(context);
     return FileEditorManagerEx.getInstanceEx(project);
   }
@@ -105,5 +163,16 @@ public class WindowGroup {
     if (editorWindow != null) {
       editorWindow.split(orientation, true, virtualFile, true);
     }
+  }
+
+  @Nullable
+  private static Rectangle getEditorWindowRectangle(@NotNull EditorWindow window) {
+    final EditorWithProviderComposite editor = window.getSelectedEditor();
+    if (editor != null) {
+      final Point point = editor.getComponent().getLocationOnScreen();
+      final Dimension dimension = editor.getComponent().getSize();
+      return new Rectangle(point, dimension);
+    }
+    return null;
   }
 }